pyproject.toml
[project]
name = "advent-of-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aocenv>=0.1.0",
]
